restyle(norma): moved concrete classes to new folder
